Split files for different compilation flags
