+++ include/decomp1.h
@@ -1,0 +1,14 @@
+#ifndef _DECOMP1_H
+#define _DECOMP1_H
+
+#include "common.h"
+
+extern s32 D_80042FB8;
+extern u8 D_80046FE0;
+extern s16 D_80042FBC;
+extern s16 D_80046FDC;
+extern s16 D_80042FBE;
+extern s16 D_80042FC0;
+extern s16 D_80042FC2;
+
+#endif
